extract dependency saving from document
